Move conjoin and omit into junebug.utils
